feat: support decimal KTW values and show warning for < 2mm
- Allow decimal input for KTW values (e.g., 1.5, 1.9)
- Display KTW < 2mm with 1 decimal place and red text color
- Show KTW warning symbol on tooth image when KTW < 2mm
- Fix editing issue by showing raw value when focused

+++ src/domain/chart/chart.image.ts
@@ -216,7 +216,8 @@ export const getKtwWarningStyle = (id: ToothId, _surface: Surface): KtwWarningPo
 }
 
 // Check if KTW warning should be displayed (KTW < 2mm and has a value)
+// Uses parseFloat to support decimal values like 1.5, 1.9
 export const shouldShowKtwWarning = (ktwValue: string): boolean => {
-  const num = parseInt(ktwValue) || 0
+  const num = parseFloat(ktwValue) || 0
   return num > 0 && num < 2
 }
